requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
cycler==0.10.0
decorator==4.4.0
Django==2.2
django-filter==2.1.0
djangorestframework==3.9.2
gunicorn==19.9.0
imageio==2.5.0
imutils==0.5.2
kiwisolver==1.0.1
Markdown==3.1
matplotlib==3.0.3
more-itertools==7.0.0
networkx==2.2
numpy==1.16.2
opencv-contrib-python==4.0.1.23
PassportEye==1.4.0
pathlib2==2.3.3
pdfminer3k==1.3.1
Pillow==6.0.0
pluggy==0.9.0
ply==3.11
py==1.8.0
pyparsing==2.4.0
pytesseract==0.2.6
pytest==4.4.0
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.2
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sqlparse==0.3.0
